+++ clipper.h
@@ -128,9 +128,9 @@ public:
      * @return number of vertices in the polygon resulting after clipping
      *
      */
-    int clipPolygon(int in, const float inx[], const float iny[],
-                    float outx[], float outy[],
-		    float x0, float y0, float x1, float y1);
+    int clipPolygon(int in, const int inx[], const int iny[],
+        int outx[], int outy[],
+        int x0, int y0, int x1, int y1);
 
 };
